# docs/swagger.yaml
definitions:
...
  dto.CreateArticleTagRequest:
    properties:
      article_id:
        type: integer
      tag_id:
        type: integer
    type: object
...
  models.ArtilceTag:
    properties:
      article_id:
        type: integer
      tag_id:
        type: integer
    type: object
...
info:
  contact: {}
paths:
  /article-tags:
    get:
      parameters:
      - description: Article ID
        in: query
        name: article_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.ArtilceTag'
            type: array
      summary: Article taglarini ko'rish
      tags:
      - ArticleTags
  /article-tags/add:
    post:
      consumes:
      - application/json
      parameters:
      - description: Article va Tag ID
        in: body
        name: article_tag
        required: true
        schema:
          $ref: '#/definitions/dto.CreateArticleTagRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Tag qo'shildi
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Articlega tag qo'shish
      tags:
      - ArticleTags
  /article-tags/delete:
    delete:
      parameters:
      - description: Article ID
        in: query
        name: article_id
        required: true
        type: integer
      - description: Tag ID
        in: query
        name: tag_id
        required: true
        type: integer
      responses:
        "200":
          description: Tag o'chirildi
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Articledan tag o'chirish
      tags:
      - ArticleTags
...
securityDefinitions:
  BearerAuth:
    description: Blog loyihasi API
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"